+++ travis/publish.sh
@@ -11,7 +11,7 @@ function deploy {
     TEMPLATE_DIR="${PWD}/rancher/${version}"
     openssl aes-256-cbc -K $encrypted_2e4df3d9a0ba_key -iv $encrypted_2e4df3d9a0ba_iv -in .rancher.env.enc -out "$HOME/.rancher.env" -d
     (. "$HOME/.rancher.env" ; cd "${TEMPLATE_DIR}" && \
-     ${RANCHER_COMPOSE} -p "${RANCHER_STACK_NAME}" rm --force && \
+     ${RANCHER_COMPOSE} -p "${RANCHER_STACK_NAME}" rm odoo db --force && \
      sleep 30 && \
      ${RANCHER_COMPOSE} -p "${RANCHER_STACK_NAME}" up --pull --recreate --force-recreate --confirm-upgrade -d)
 }